Remove InlineeManager.
It wasn't doing anything useful, removing it doesn't change the output
on the files I've tested.

`symbolic::debuginfo::Function` has already normalized the line info for
us so that all the lines of the outer function describe locations in the
outer function. Normalizing was only needed when we were manually using
DWARF info, which flattens lines from all levels of the inline nesting
into one list. But symbolic already undoes that flattening for us: Lines
inside inlined functions are stored on the `Function` for the inline
function, not on the `Function` for the outer function. The outer
function only has the call location for the inlines in its line info,
which is exactly what we need.