Implement --inlines for ELF and mach-O.
When dump_syms is run with --inlines, we now emit INLINE_ORIGIN and INLINE directives.

These allow recovering the inline stack at each address.

This is currently only implemented for ELF / mach-O, not for pdb.

When dump_syms is run in this mode, it also outputs the "line" records
slightly differently: For addresses which contain an inlined call, rather
than describing the line number in the outer function at that call, the
line record now describes the file and line number of the innermost
inlined call at that address.
Furthermore, we no longer filter out line numbers that are zero.
Compilers emit zero line numbers to say "I lost track of the original
line which contributed this instruction". Without --inlines, dump_syms
skips zero lines and takes the line number from the previous record. But
this line number can often be wrong, and it seems better to preserve the
information that there is no line number. So with --inlines we now output
zero line numbers whenever the compiler produced debug info with zero
line numbers.

Consumers of this format need to be updated to understand zero line
numbers (as "line number is unknown") in .sym files which contain INLINE
information.