Don't strip underscores from pdb FUNC names.
When we have a procedure symbol, and then later find a public symbol
at the same address, we try to get the parameter size from the public
symbol's name. But we didn't only got the parameter size, we also
overwrote the function name with the undecorated function name.
This overwriting is not necessary:

 - If the function has a non-zero TypeIndex, we will use the TypeFormatter
   to add parameters, and we want to add those to the original function
   name, not to the undecorated name from the PUBLIC.
 - If the function has a zero TypeIndex, we will call FuncName::get_unknown
   again, and it will get undecorated at that point.

Storing the undecorated name here meant that some functions lost their
leading underscores.