Always substitute parameterless functions with the symbol name.
Even if the symbol name is not MSVC-mangled, it's often a better choice
for the function name. In the example I was looking at, the function
name started with an underscore, and taking the name from the symbol
stripped that underscore, which looked better.
We don't strip underscores from function names normally, because
function names are already expected to be clean.