Add more sensible way for constructing dumper::Config
It is rather cumbersome to create a dumper::Config object, owing to the
many fields that conceptually have sensible defaults but have to be
named explicitly nevertheless.
This change introduces a new constructor that instantiates an instance
given only an Output object as input -- all other fields are set to
sensible defaults. This is also a good first step towards making the
type less prone to semver violations caused by additions, because it
more easily enables making it non-exhaustive (which really should be
done, but isn't as part of this change, because it would be a semver
breaking change in and of itself).